{{cookiecutter.project_slug}}/requirements.txt
asgiref==3.3.1
Babel==2.9.0
Django==3.0.11
django-environ==0.4.5
django-extra-views==0.13.0
django-haystack==3.0
django-oscar==2.1
django-phonenumber-field==3.0.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==4.17.1
phonenumbers==8.12.13
Pillow==8.0.1
purl==1.5
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
